renamed users entity
