+++ src/typeorm/index.ts
@@ -1,4 +1,4 @@
-import { Users } from "./Entities/User";
+import { Users } from "./Entities/Users";
 const entities = [Users];
 export {Users};
 export default entities;
